+++ src/App.js
@@ -2,14 +2,14 @@ import logo from './logo.svg';
 import './App.css';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faReact } from '@fortawesome/free-brands-svg-icons';
-import About from "./componeents/About";
-import Header from "./componeents/Header";
+import Header from "./components/Header";
+import Home from './components/Home';
 
 function App() {
   return (
     <div className='relative bg-fuchsia-500'>
       <Header />
-      <About />
+      <Home />
     </div>
   );
 }
